Moved IfcImporterUtilT to separate file to enable compilation.
